requirements.txt
matplotlib
numpy
Pillow
rasterio
scikit_learn
seaborn
kornia